+++ contracts/interfaces/lotterySwap.sol
@@ -1,28 +1,25 @@
 pragma solidity ^0.6.2;
 
 interface LotterySwapInterface {
-    function getEthToDaiProfit(uint256 totalProfit) external view returns (uint256 requiredAmt);
-
-    function getTokenToDaiProfit(uint256 totalProfit) external view returns (uint256 requiredAmt);
-
-    function getEthToDaiFee(uint256 totalAmt) external view returns (uint256 requiredAmt);
-
-    function getTokenToDaifee(uint256 totalAmt) external view returns (uint256 requiredAmt);
+    function getEthToDaiProfit(uint totalProfit) external view returns(uint requiredAmt);
+    function getTokenToDaiProfit(uint totalProfit) external view returns(uint requiredAmt);
+    function getEthToDaiFee(uint totalAmt) external view returns(uint requiredAmt);
+    function getTokenToDaifee(uint totalAmt) external view returns(uint requiredAmt);
 
     function swapEthToDai(
         address payable user,
         address hibachiFor,
-        uint256 totalAmt,
+        uint totalAmt,
         bool isFee,
         bool isIn
-    ) external payable returns (uint256 leftAmt, uint256 candies);
+    ) external payable returns(uint leftAmt, uint candies);
 
     function swapTokenToDai(
         address user,
         address hibachiFor,
         address token,
-        uint256 totalAmt,
+        uint totalAmt,
         bool isFee,
         bool isIn
-    ) external returns (uint256 leftAmt, uint256 candies);
+    ) external returns(uint leftAmt, uint candies);
 }
